requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
numpy==1.24.3
scipy==1.11.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0

